product, services and product category updated

+++ Petshop/ProductCategory.cs
@@ -31,13 +31,14 @@ namespace Petshop
             dbConnect = new Conclass();
             dbConnect.OpenConnection();
             DataTable dt = new DataTable();
-            mySqlDataAdapter = new MySqlDataAdapter("SELECT * FROM productcategory", dbConnect.myconnect);
+            mySqlDataAdapter = new MySqlDataAdapter("SELECT * FROM productcategory WHERE pCategory_name != 'NULL'", dbConnect.myconnect);
             mySqlDataAdapter.Fill(dt);
             pCategory.DataSource = dt;
             pCategory.Columns[0].Visible = true;
             DataGridViewColumn column1 = pCategory.Columns["pCategory_name"];
             pCategory.Columns[0].HeaderText = "Category ID";
             pCategory.Columns[1].HeaderText = "Product Category";
+            dbConnect.CloseConnection();
           
         }
 
@@ -54,16 +55,35 @@ namespace Petshop
             }
             dbConnect = new Conclass();
             dbConnect.OpenConnection();
-            cmd = new MySqlCommand("INSERT INTO productcategory (pCategory_name) VAlUES (@pCname)", dbConnect.myconnect);
-            cmd.Parameters.AddWithValue("@pCname", pCategoryName.Text);
-            cmd.ExecuteNonQuery();
-            MaterialMessageBox.Show("New Product Category Successfully Added", "Success");
-            pCategoryName.Clear();
-            populategv();
+            cmd = new MySqlCommand("SELECT * FROM productcategory WHERE pCategory_name = @Cname", dbConnect.myconnect);
+            cmd.Parameters.AddWithValue("@Cname", pCategoryName.Text);
+            myReader = cmd.ExecuteReader();
+            if(myReader.Read() == true)
+            {
+                MaterialMessageBox.Show("Product Category Already exist", "Notice", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                dbConnect.CloseConnection();
+                return;
+            }
+            else
+            {
+                dbConnect = new Conclass();
+                dbConnect.OpenConnection();
+                cmd = new MySqlCommand("INSERT INTO productcategory (pCategory_name) VAlUES (@pCname)", dbConnect.myconnect);
+                cmd.Parameters.AddWithValue("@pCname", pCategoryName.Text);
+                cmd.ExecuteNonQuery();
+                MaterialMessageBox.Show("New Product Category Successfully Added", "Success");
+                pCategoryName.Clear();
+                populategv();
+                dbConnect.CloseConnection();
+            }
+
+
+          
         }
 
         private void Edit_Click(object sender, EventArgs e)
         {
+
             dbConnect = new Conclass();
             dbConnect.OpenConnection();
             cmd = new MySqlCommand("UPDATE productcategory SET pCategory_name = @Cname WHERE pCategory_id = @uid", dbConnect.myconnect);
@@ -73,24 +93,46 @@ namespace Petshop
             MaterialMessageBox.Show("Updated Successfully", "Success");
             pCategoryName.Clear();
             populategv();
+            dbConnect.CloseConnection();
         }
 
         private void button2_Click(object sender, EventArgs e)
         {
-            dbConnect = new Conclass();
-            dbConnect.OpenConnection();
-            cmd = new MySqlCommand("DELETE FROM productcategory WHERE pCategory_id = @did", dbConnect.myconnect);
-            cmd.Parameters.AddWithValue("@did", CategoryID.Text);
-            cmd.ExecuteNonQuery();
-            MaterialMessageBox.Show("Deleted Successfully", "Success");
-            pCategoryName.Clear();
-            populategv();
+            if (MaterialMessageBox.Show("Are you sure you want to update this product?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            {
+                dbConnect = new Conclass();
+                dbConnect.OpenConnection();
+                cmd = new MySqlCommand("UPDATE product SET pCategory_id = 8 WHERE pCategory_id = @did", dbConnect.myconnect);
+                cmd.Parameters.AddWithValue("@did", CategoryID.Text);
+                cmd.ExecuteNonQuery();
+                dbConnect = new Conclass();
+                dbConnect.OpenConnection();
+                cmd = new MySqlCommand("DELETE FROM productcategory WHERE pCategory_id = @did", dbConnect.myconnect);
+                cmd.Parameters.AddWithValue("@did", CategoryID.Text);
+                cmd.ExecuteNonQuery();
+                MaterialMessageBox.Show("Deleted Successfully", "Success");
+                pCategoryName.Clear();
+                populategv();
+                dbConnect.CloseConnection();
+            }
         }
 
         private void pCategory_MouseClick(object sender, MouseEventArgs e)
         {
-            CategoryID.Text = pCategory.SelectedRows[0].Cells[0].Value.ToString();
-            pCategoryName.Text = pCategory.SelectedRows[0].Cells[1].Value.ToString();
+          
+        }
+
+        private void pCategory_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+            if (e.RowIndex < 0)
+            {
+
+            }
+            else
+            {
+                CategoryID.Text = pCategory.Rows[e.RowIndex].Cells[0].Value.ToString();
+                pCategoryName.Text = pCategory.Rows[e.RowIndex].Cells[1].Value.ToString();
+            }
         }
     }
 }
